Rework code from the common/ directory to use correct types for variables:  - use uint32_t for RTP TimeStamp (TS), TS_SCALED and TS_STRIDE,  - use uint16_t for SN and inner/outer IP-IDs,  - use size_t for all lengths and sizes.
To achieve this objective, the following modifications were required:
 - Change the prototypes of several functions that used the returned value
   for both the status code and the result (eg. -1 for error, > 0 for result).
   This was not possible anymore with result stored in unsigned variables.
 - Use the const keyword as much as possible in the new prototypes when the
   input parameters are not modified (and should not be modified).
 - Rewrite some bitmask computation to avoid integer overflow.
 - Add many assertion statements to ease debugging.
 - Move the f() function in separate files interval.[ch] because the function
   is shared between wlsb and lsb modules.

+++ src/comp/c_generic.h
@@ -30,6 +30,7 @@
 #include "comp_list.h"
 #include "ip.h"
 
+#include <stdlib.h>
 #include <netinet/ip.h>
 
 
@@ -84,7 +85,8 @@ struct ipv4_header_info
 	int old_nbo;
 
 	/// The delta between the IP-ID and the current Sequence Number (SN)
-	int id_delta;
+	/// (overflow over 16 bits is expected when SN > IP-ID)
+	uint16_t id_delta;
 };
 
 
@@ -155,11 +157,11 @@ struct generic_tmp_variables
 	int send_dynamic;
 
 	/// The number of bits needed to encode the Sequence Number (SN)
-	int nr_sn_bits;
+	size_t nr_sn_bits;
 	/// The number of bits needed to encode the IP-ID of the outer IP header
-	int nr_ip_id_bits;
+	size_t nr_ip_id_bits;
 	/// The number of bits needed to encode the IP-ID of the inner IP header
-	int nr_ip_id_bits2;
+	size_t nr_ip_id_bits2;
 
 	/// The type of packet the compressor must send: IR, IR-DYN, UO*
 	rohc_packet_t packet_type;
